% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_adj_matrix.R
\name{get_adj_matrix}
\alias{get_adj_matrix}
\title{Set up an adjadency matrix}
\usage{
get_adj_matrix(user_list, mention_list)
}
\arguments{
\item{user_list}{chr-vector or -list containing multiple Twitter user names}

\item{mention_list}{a list of all interactions - e.g. following, retweet, mention - of one users with other users}
}
\value{
Returns an matrix containing the number of directed interactions from a user -cols- to another user -rows-
}
\description{
This function sets up an adjacency matrix from a
list of interactions between Twitter users. This function can be used to set up
a mention-, following- or retweet-matrix.
}
\examples{
get_adj_matrix(user_list, retweet_list)
}

