% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_user_list.R
\name{update_user_list}
\alias{update_user_list}
\title{Update a twitter user list}
\usage{
update_user_list(user_list)
}
\arguments{
\item{user_list}{chr-vector or -list containing multiple Twitter user names}
}
\value{
Returns an updated user list without the inactive Twitter accounts
}
\description{
Fix for an errorness list of Twitter users. Also prints names of inactive accounts.
}
\examples{
update_user_list(c("RStudio", "hadleywickham"))
}

